Rename foldend server module to handler_server
