+++ src/foldend/handler_server/endpoints/trace_handler_stream.rs
