+++ src/foldend/handler_server/mod.rs
@@ -1,0 +1,3 @@
+mod endpoints;
+mod handler_service;
+pub mod server;
